Added reservepercentage and claimreferrals option
Reservepercentage allows you to keep a percentage of your funds in your trading wallet, so you can easier claim them.
Be aware that when you have a big percentage in here, it will NOT be used for trading and thus will not help building your returns further!

Claimreferrals 1 or 0, if set to 1 we will visit the referral claim page every cycle. If you couldn't care less abut claiming referrels, you can turn it off to save some time.